# manager_api_getaway/api/docs/swagger.yaml
definitions:
  manager_service.CreateGroupRequest:
    properties:
      branch:
        type: string
      group_name:
        type: string
      journal_id:
        type: string
      support_teacher:
        type: string
      teacher:
        type: string
      type_id:
        type: string
    type: object
  manager_service.CreateJournalRequest:
    properties:
      from_date:
        type: string
      schedules_id:
        type: string
      student_name:
        type: string
      to_date:
        type: string
    type: object
  manager_service.CreateLessonRequest:
    properties:
      lesson_name:
        type: string
      task_id:
        type: string
    type: object
...
  manager_service.CreateSchedulesRequest:
    properties:
      end_time:
        type: string
      lesson_id:
        type: string
      start_time:
        type: string
    type: object
  manager_service.CreateTaskRequest:
    properties:
      created_by:
        type: string
      deadline:
        type: string
      label:
        type: string
      score:
        type: number
    type: object
  manager_service.CreateTypeRequest:
    properties:
      course_name:
        type: string
      description:
        type: string
    type: object
  manager_service.Journal:
    properties:
      from_date:
        type: string
      id:
        type: string
      schedules:
        $ref: '#/definitions/manager_service.Schedules'
      schedules_id:
        type: string
      student_name:
        type: string
      students:
        items:
          $ref: '#/definitions/manager_service.Student'
        type: array
      to_date:
        type: string
    type: object
  manager_service.Lesson:
    properties:
      id:
        type: string
      lesson_name:
        type: string
      task:
        $ref: '#/definitions/manager_service.Task'
      task_id:
        type: string
    type: object
  manager_service.Schedules:
    properties:
      end_time:
        type: string
      id:
        type: string
      lesoon:
        $ref: '#/definitions/manager_service.Lesson'
      lesson_id:
        type: string
      start_time:
        type: string
    type: object
  manager_service.Student:
    properties:
      first_name:
        type: string
      id:
        type: string
      last_name:
        type: string
    type: object
  manager_service.Task:
    properties:
      created_by:
        type: string
      deadline:
        type: string
      id:
        type: string
      label:
        type: string
      score:
        type: number
    type: object
  manager_service.TaskScore:
    properties:
      id:
        type: string
      score:
        type: string
      student_id:
        type: string
      task_id:
        type: string
    type: object
  manager_service.TeacherGetTeacherInfoResponse:
    properties:
      journals:
        items:
          $ref: '#/definitions/manager_service.Journal'
        type: array
      schedules:
        items:
          $ref: '#/definitions/manager_service.Schedules'
        type: array
      students:
        items:
          $ref: '#/definitions/manager_service.Student'
        type: array
      task:
        items:
          $ref: '#/definitions/manager_service.Task'
        type: array
      tasl_score:
        items:
          $ref: '#/definitions/manager_service.TaskScore'
        type: array
      teacher:
        $ref: '#/definitions/manager_service.TeacherTeacher'
    type: object
  manager_service.TeacherTeacher:
    properties:
      branch:
        type: string
      fullname:
        type: string
      group_id:
        type: integer
      id:
        type: string
      ielts_attempts_count:
        type: integer
      ielts_score:
        type: number
      password:
        type: string
      phone:
        type: string
      salary:
        type: number
    type: object
...
  models.ResponseError:
    properties:
      error: {}
    type: object
  models.ResponseSuccess:
    properties:
      data: {}
      metadata: {}
    type: object
...
info:
  contact: {}
paths:
...
  /forsupportteacher/{id}:
    get:
      consumes:
      - application/json
      description: Retrieves a forsupportteacher by its ID
      parameters:
      - description: forsupportteacher ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      security:
      - ApiKeyAuth: []
      summary: Get a forsupportteacher by its ID
      tags:
      - forsupportteacher
  /forteacher/{id}:
    get:
      consumes:
      - application/json
      description: Retrieves a forteacher by its ID
      parameters:
      - description: forteacher ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/manager_service.TeacherGetTeacherInfoResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      security:
      - ApiKeyAuth: []
      summary: Get a forteacher by its ID
      tags:
      - forteacher
  /groups:
    post:
      consumes:
      - application/json
      description: Create a new group
      parameters:
      - description: group
        in: body
        name: group
        required: true
        schema:
          $ref: '#/definitions/manager_service.CreateGroupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      security:
      - ApiKeyAuth: []
      summary: Create a group
      tags:
      - groups
  /groups/{id}:
    get:
      consumes:
      - application/json
      description: Retrieves a group by its ID
      parameters:
      - description: group ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      security:
      - ApiKeyAuth: []
      summary: Get a group by its ID
      tags:
      - groups
  /journal:
    post:
      consumes:
      - application/json
      description: Create a new journal
      parameters:
      - description: journal
        in: body
        name: group
        required: true
        schema:
          $ref: '#/definitions/manager_service.CreateJournalRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      security:
      - ApiKeyAuth: []
      summary: Create a journal
      tags:
      - journal
  /journal/{id}:
    get:
      consumes:
      - application/json
      description: Retrieves a journal by its ID
      parameters:
      - description: journal ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      security:
      - ApiKeyAuth: []
      summary: Get a journal by its ID
      tags:
      - journal
  /lesson:
    post:
      consumes:
      - application/json
      description: Create a new lesson
      parameters:
      - description: talessonsk
        in: body
        name: group
        required: true
        schema:
          $ref: '#/definitions/manager_service.CreateLessonRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      security:
      - ApiKeyAuth: []
      summary: Create a lesson
      tags:
      - lesson
  /lesson/{id}:
    get:
      consumes:
      - application/json
      description: Retrieves a lesson by its ID
      parameters:
      - description: lesson ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      security:
      - ApiKeyAuth: []
      summary: Get a lesson by its ID
      tags:
      - lesson
...
  /schedule:
    post:
      consumes:
      - application/json
      description: Create a new schedule
      parameters:
      - description: 'schedule '
        in: body
        name: group
        required: true
        schema:
          $ref: '#/definitions/manager_service.CreateSchedulesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      security:
      - ApiKeyAuth: []
      summary: Create a schedule
      tags:
      - schedule
  /schedule/{id}:
    get:
      consumes:
      - application/json
      description: Retrieves a schedule by its ID
      parameters:
      - description: schedule ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      security:
      - ApiKeyAuth: []
      summary: Get a schedule by its ID
      tags:
      - schedule
...
  /task:
    post:
      consumes:
      - application/json
      description: Create a new task
      parameters:
      - description: task
        in: body
        name: group
        required: true
        schema:
          $ref: '#/definitions/manager_service.CreateTaskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      security:
      - ApiKeyAuth: []
      summary: Create a task
      tags:
      - task
  /task/{id}:
    get:
      consumes:
      - application/json
      description: Retrieves a task by its ID
      parameters:
      - description: task ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      security:
      - ApiKeyAuth: []
      summary: Get a task by its ID
      tags:
      - task
...
  /type:
    post:
      consumes:
      - application/json
      description: Create a new type
      parameters:
      - description: 'type '
        in: body
        name: group
        required: true
        schema:
          $ref: '#/definitions/manager_service.CreateTypeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      security:
      - ApiKeyAuth: []
      summary: Create a type
      tags:
      - type
  /type/{id}:
    get:
      consumes:
      - application/json
      description: Retrieves a type by its ID
      parameters:
      - description: type ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      security:
      - ApiKeyAuth: []
      summary: Get a type by its ID
      tags:
      - type
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"